requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
pyyaml>=6.0
pycocotools>=2.0.6
tqdm>=4.65.0
nltk>=3.8.1
numpy>=1.24.0
torchvision>=0.15.0 